CI fixes (RPI)

+++ Svc/TlmPacketizer/TlmPacketizer.cpp
@@ -84,7 +84,7 @@ void TlmPacketizer::setPacketList(const TlmPacketizerPacketList& packetList,
         }  // end channel in packet
         FW_ASSERT(packetLen <= FW_COM_BUFFER_MAX_SIZE, static_cast<FwAssertArgType>(packetLen), static_cast<FwAssertArgType>(pktEntry));
         // clear contents
-        memset(this->m_fillBuffers[pktEntry].buffer.getBuffAddr(), 0, packetLen);
+        memset(this->m_fillBuffers[pktEntry].buffer.getBuffAddr(), 0, static_cast<size_t>(packetLen));
         // serialize packet descriptor and packet ID now since it will always be the same
         Fw::SerializeStatus stat = this->m_fillBuffers[pktEntry].buffer.serialize(
             static_cast<FwPacketDescriptorType>(Fw::ComPacket::FW_PACKET_PACKETIZED_TLM));
